// src/models/vans/law.ts
export interface createVan {

    uuid: string;
    make: string;
    reg_no: string;
    type: string;
    capacity: string;
    provider: string;
    gps_tracking_id: string;
    documents:string;
    dc_uuid: string;
    dc_name: string;
    color_code: string;
    status:string;
    created_on: number;
    updated_on: number;
    
}